+++ src/krill/classes/prepare.d/welcome.tsx
@@ -14,7 +14,7 @@ import Welcome from '../../components/welcome.js'
 import Prepare from '../prepare.js'
 import Utils from '../../../classes/utils.js'
 import selectLanguages from '../../lib/select_languages.js'
-import { IWelcome } from '../../interfaces/i-krill.js'
+import { IWelcome } from '../../interfaces/i_krill.js'
 
 
 /**
