+++ src/krill/classes/sequence.d/bootloader_config.ts
@@ -10,7 +10,7 @@
 import Openmamba from '../../../classes/pacman.d/openmamba.js'
 import Utils from '../../../classes/utils.js'
 import { exec } from '../../../lib/utils.js'
-import Sequence from '../../classes/sequence.js'
+import Sequence from '../sequence.js'
 
 /**
  *
