+++ src/krill/classes/sequence.d/del_live_user.ts
@@ -9,7 +9,7 @@
 
 import Utils from '../../../classes/utils.js'
 import { exec } from '../../../lib/utils.js'
-import Sequence from '../../classes/sequence.js'
+import Sequence from '../sequence.js'
 
 /**
  * delUser
