+++ src/krill/classes/sequence.d/fstab.ts
@@ -8,12 +8,12 @@
  */
 
 import shx from 'shelljs'
-import { InstallationMode } from '../../classes/krill-enums.js'
+import { InstallationMode } from '../krill_enums.js'
 
 import Pacman from '../../../classes/pacman.js'
 import Utils from '../../../classes/utils.js'
 import Sequence from '../../classes/sequence.js'
-import { SwapChoice } from '../../classes/krill-enums.js'
+import { SwapChoice } from '../krill_enums.js'
 
 /**
  * fstab()
