+++ src/krill/classes/sequence.d/grubcfg.ts
@@ -11,8 +11,8 @@ import fs from 'node:fs'
 
 import Utils from '../../../classes/utils.js'
 import Sequence from '../../classes/sequence.js'
-import { SwapChoice } from '../../classes/krill-enums.js'
-import { InstallationMode } from '../../classes/krill-enums.js'
+import { SwapChoice } from '../krill_enums.js'
+import { InstallationMode } from '../krill_enums.js'
 /**
  * grubcfg
  * - open /etc/default/grub
