+++ src/krill/classes/sequence.d/m_keyboard.ts
@@ -11,7 +11,7 @@ import fs from 'node:fs'
 
 import Utils from '../../../classes/utils.js'
 import { exec } from '../../../lib/utils.js'
-import Sequence from '../../classes/sequence.js'
+import Sequence from '../sequence.js'
 
 /**
  *
