+++ src/krill/classes/sequence.d/machine_id.ts
@@ -11,7 +11,7 @@
 import fs from 'node:fs'
 
 import { exec } from '../../../lib/utils.js'
-import Sequence from '../../classes/sequence.js'
+import Sequence from '../sequence.js'
 import Utils from '../../../classes/utils.js'
 
 /**
