+++ src/krill/classes/sequence.d/mount_fs.ts
@@ -10,8 +10,8 @@
 import fs from 'node:fs'
 
 import { exec } from '../../../lib/utils.js'
-import Sequence from '../../classes/sequence.js'
-import { InstallationMode, SwapChoice } from '../../classes/krill-enums.js'
+import Sequence from '../sequence.js'
+import { InstallationMode, SwapChoice } from '../krill_enums.js'
 import Utils from '../../../classes/utils.js'
 
 /**
