+++ src/krill/classes/sequence.d/mount_vfs.ts
@@ -8,7 +8,7 @@
  */
 
 import { exec } from '../../../lib/utils.js'
-import Sequence from '../../classes/sequence.js'
+import Sequence from '../sequence.js'
 /**
  * mountvfs()
  */
