+++ src/krill/classes/sequence.d/remove_installer_link.ts
@@ -9,7 +9,7 @@
 
 import fs from 'node:fs'
 
-import Sequence from '../../classes/sequence.js'
+import Sequence from '../sequence.js'
 
 /**
  * removeInstallerLink
