+++ src/krill/classes/sequence.tsx
@@ -33,41 +33,41 @@ import Utils from '../../classes/utils.js'
 import Xdg from '../../classes/xdg.js';
 
 import { IInstaller, IDevices, IDevice } from '../../interfaces/index.js'
-import { ICalamaresModule, ILocation, IKeyboard, IPartitions, IUsers } from '../interfaces/i-krill.js'
+import { ICalamaresModule, ILocation, IKeyboard, IPartitions, IUsers } from '../interfaces/i_krill.js'
 import { exec } from '../../lib/utils.js'
 
 // import krill modules
 import partition from './sequence.d/partition.js'
 // import createLvmPartitions from './sequence.d/partition.d/create-lvm-partitions.js'
-import biosStandard from './sequence.d/partition.d/bios-standard.js'
-import biosLuks from './sequence.d/partition.d/bios-luks.js'
-import uefiStandard from './sequence.d/partition.d/uefi-standard.js'
-import uefiLuks from './sequence.d/partition.d/uefi-luks.js'
+import biosStandard from './sequence.d/partition.d/bios_standard.js'
+import biosLuks from './sequence.d/partition.d/bios_luks.js'
+import uefiStandard from './sequence.d/partition.d/uefi_standard.js'
+import uefiLuks from './sequence.d/partition.d/uefi_luks.js'
 
-import { mountFs, umountFs } from './sequence.d/mount-fs.js'
-import { mountVfs, umountVfs } from './sequence.d/mount-vfs.js'
+import { mountFs, umountFs } from './sequence.d/mount_fs.js'
+import { mountVfs, umountVfs } from './sequence.d/mount_vfs.js'
 import unpackfs from './sequence.d/unpackfs.js'
-import machineId from './sequence.d/machine-id.js'
+import machineId from './sequence.d/machine_id.js'
 import fstab from './sequence.d/fstab.js'
 import locale from './sequence.d/locale.js'
-import mKeyboard from './sequence.d/m-keyboard.js'
-import localeCfg from './sequence.d/locale-cfg.js'
+import mKeyboard from './sequence.d/m_keyboard.js'
+import localeCfg from './sequence.d/locale_cfg.js'
 // users
-import addUser from './sequence.d/add-user.js'
-import changePassword from './sequence.d/change-password.js'
+import addUser from './sequence.d/add_user.js'
+import changePassword from './sequence.d/change_password.js'
 // displaymanager: autologin
-import networkCfg from './sequence.d/network-cfg.js'
+import networkCfg from './sequence.d/network_cfg.js'
 // hwclock:
 // services-systemd:
 // bootloader-config
-import bootloaderConfig from './sequence.d/bootloader-config.js'
+import bootloaderConfig from './sequence.d/bootloader_config.js'
 import grubcfg from './sequence.d/grubcfg.js'
 import bootloader from './sequence.d/bootloader.js'
 import packages from './sequence.d/packages.js'
-import removeInstallerLink from './sequence.d/remove-installer-link.js'
-import initramfsCfg from './sequence.d/initramfs-cfg.js'
+import removeInstallerLink from './sequence.d/remove_installer_link.js'
+import initramfsCfg from './sequence.d/initramfs_cfg.js'
 import initramfs from './sequence.d/initramfs.js'
-import delLiveUser from './sequence.d/del-live-user.js'
+import delLiveUser from './sequence.d/del_live_user.js'
 import umount from './sequence.d/umount.js'
 import mkfs from './sequence.d/mkfs.js'
 import hostname from './sequence.d/hostname.js'
