+++ src/krill/interfaces/i_krill.ts
@@ -6,7 +6,7 @@
  * license: MIT
  */
 
-import { SwapChoice } from '../classes/krill-enums.js'
+import { SwapChoice } from '../classes/krill_enums.js'
 
 export interface IWelcome {
   language: string
