+++ src/krill/lib/select_installation_mode.ts
@@ -7,7 +7,7 @@
  */
 
 import inquirer from 'inquirer'
-import { InstallationMode } from '../classes/krill-enums.js'
+import { InstallationMode } from '../classes/krill_enums.js'
 
 export default async function selectInstallationMode(): Promise<InstallationMode> {
   const modes = Object.values(InstallationMode)
